+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0
+### Added
+- `BitSetInterface::is_empty`.
+- All `BitSetInterface`s now implement `Eq`.
+- `BitSet` (without &) now implements op's too.
+
 ## 0.2.0
 ### Changed
 - `IConfig` renamed to `Config`.